feat: enum for http endpoints refactor: enum Type moved to enums package
