+++ src/util/enums/Endpoint.java
@@ -1,0 +1,13 @@
+package util.enums;
+
+public enum Endpoint {
+    INVALID_SUBRESOURCE,
+
+    INVALID,
+    INVALID_METHOD,
+
+    OPTIONS,
+    GET,
+    POST,
+    DELETE
+}
