+++ src/util/enums/Type.java
@@ -1,4 +1,4 @@
-package util;
+package util.enums;
 
 public enum Type {
     TASK,
